Remove AbstractName
